+++ CHANGELOG.md
@@ -5,10 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+:warning: Major version! Following the migration of the "Calypsonet Terminal" APIs to the
+[Eclipse Keypop project](https://keypop.org), this library now implements Keypop interfaces.
 ### Added
 - Added project status badges on `README.md` file.
 ### Fixed
 - CI: code coverage report when releasing.
+### Upgraded
+- "Keyple Service Resource Library" to version `3.0.0`
+- "Keyple Util Library" to version `2.3.1`
 
 ## [1.0.1] - 2023-04-27
 ### Fixed